+++ packages/shared/src/catalog/index.ts
@@ -1,0 +1,4 @@
+export * from "./types";
+export * from "./outputs/toRacProps";
+export * from "./primitives/button";
+export * from "./registry";